rename validators.py to directories.py
