+++ jetbase/config.py
@@ -8,7 +8,7 @@ from typing import Any
 
 import tomli
 
-from jetbase.commands.validators import get_jetbase_directory
+from jetbase.commands.directories import get_jetbase_directory
 from jetbase.constants import ENV_FILE
